requirements.txt
torch==2.2.2
torchvision==0.17.2
opencv-python==4.9.0.80
scipy==1.13.0
scikit-learn==1.4.2
matplotlib==3.8.4
tqdm==4.66.4
numpy==1.26.4
Pillow==10.3.0
h5py==3.11.0
seaborn==0.13.2
ipywidgets==7.8.5
jupyterlab
